// eAccountingClient/src/app/models/cashRegister.model.ts
import { CashRegisterDetailModel } from './cashRegisterDetail.model';
import { CurrencyTypeModel } from './currencyType.model';

export class CashRegisterModel {
  id: string = '';
  name: string = '';
  depositAmount: number = 0;
  withdrawalAmount: number = 0;
  currencyType: CurrencyTypeModel = new CurrencyTypeModel();
  currencyTypeValue: number = 1;
  details: CashRegisterDetailModel[] = [];
}